+++ querybook/webapp/components/AIAssistant/AICommandBar.tsx
@@ -2,20 +2,21 @@ import { uniq } from 'lodash';
 import React, { forwardRef, useCallback, useEffect, useState } from 'react';
 
 import { AICommandInput } from 'components/AIAssistant/AICommandInput';
-import { TableSelector } from 'components/AIAssistant/TableSelector';
-import { QueryComparison } from 'components/TranspileQueryModal/QueryComparison';
+import { AICommandResultView } from 'components/AIAssistant/AICommandResultView';
 import { ComponentType, ElementType } from 'const/analytics';
-import { IQueryCellCommand, QUERY_CELL_COMMANDS } from 'const/command';
+import {
+    IQueryCellCommand,
+    QUERY_CELL_COMMANDS,
+    QueryCellCommandType,
+} from 'const/command';
 import { IQueryEngine } from 'const/queryEngine';
 import { CommandRunner, useCommand } from 'hooks/useCommand';
 import { useEvent } from 'hooks/useEvent';
 import { useForwardedRef } from 'hooks/useForwardedRef';
-import useNonEmptyState from 'hooks/useNonEmptyState';
 import { trackClick } from 'lib/analytics';
 import { TableToken } from 'lib/sql-helper/sql-lexer';
 import { matchKeyPress } from 'lib/utils/keyboard';
 import { analyzeCode } from 'lib/web-worker';
-import { Button } from 'ui/Button/Button';
 import { Message } from 'ui/Message/Message';
 import { IResizableTextareaHandles } from 'ui/ResizableTextArea/ResizableTextArea';
 
@@ -31,20 +32,6 @@ interface IQueryCellCommandBarProps {
     onUpdateEngineId: (engineId: number) => void;
     onFormatQuery: () => void;
     ref: React.Ref<IResizableTextareaHandles>;
-}
-
-interface IAICommandResultViewProps {
-    command: IQueryCellCommand;
-    commandKwargs: Record<string, any>;
-    metastoreId: number;
-    originalQuery: string;
-    tables: string[];
-    commandResult: Record<string, any>;
-    isStreaming: boolean;
-    onContinue: () => void;
-    onTablesChange: (tables: string[]) => void;
-    onAccept: (query: string) => void;
-    onDiscard: () => void;
 }
 
 const useTablesInQuery = (query: string, language: string) => {
@@ -69,125 +56,6 @@ const useTablesInQuery = (query: string, language: string) => {
     return tables;
 };
 
-const AICommandResultView = ({
-    command,
-    commandKwargs,
-    metastoreId,
-    originalQuery,
-    tables,
-    commandResult,
-    isStreaming,
-    onContinue,
-    onTablesChange,
-    onAccept,
-    onDiscard,
-}: IAICommandResultViewProps) => {
-    const [newQuery, setNewQuery] = useNonEmptyState<string>('');
-    const [explanation, setExplanation] = useState<string>('');
-    const [foundTables, setFoundTables] = useState<boolean>(false);
-
-    useEffect(() => {
-        const { type, data } = commandResult;
-
-        if (type === 'tables') {
-            onTablesChange(data);
-            setFoundTables(true);
-        } else {
-            const {
-                explanation,
-                query: rawNewQuery,
-                data: additionalData,
-            } = data;
-            setExplanation(explanation || additionalData);
-            setNewQuery(rawNewQuery);
-            setFoundTables(false);
-        }
-    }, [commandResult]);
-
-    const handleAccept = useCallback(() => {
-        onAccept(newQuery);
-        trackClick({
-            component: ComponentType.AI_ASSISTANT,
-            element: ElementType.QUERY_GENERATION_APPLY_BUTTON,
-            aux: {
-                mode: command.name,
-                question: commandKwargs.question,
-                tables,
-                query: newQuery,
-            },
-        });
-    }, [onAccept, newQuery]);
-
-    const handleDiscard = useCallback(() => {
-        onDiscard();
-        if (newQuery) {
-            trackClick({
-                component: ComponentType.AI_ASSISTANT,
-                element: ElementType.QUERY_GENERATION_REJECT_BUTTON,
-                aux: {
-                    mode: command.name,
-                    question: commandKwargs.question,
-                    tables,
-                    query: newQuery,
-                },
-            });
-        }
-    }, [onAccept, newQuery]);
-
-    const tablesDOM = (
-        <div className="mt12">
-            <div>Please review table(s) to use for the query</div>
-            <TableSelector
-                metastoreId={metastoreId}
-                tableNames={tables}
-                onTableNamesChange={(tables) => {
-                    onTablesChange(tables);
-                    setFoundTables(true);
-                }}
-            />
-            {foundTables && tables.length > 0 && (
-                <div className="mt12">
-                    <Button
-                        title="Continue"
-                        onClick={onContinue}
-                        color="accent"
-                    />
-                </div>
-            )}
-        </div>
-    );
-
-    const queryDiffDOM = (originalQuery || newQuery) && (
-        <div className="mt12">
-            <QueryComparison
-                fromQuery={originalQuery}
-                toQuery={newQuery}
-                fromQueryTitle="Original Query"
-                toQueryTitle="New Query"
-                disableHighlight={isStreaming}
-                hideEmptyQuery={true}
-                autoHeight={true}
-            />
-        </div>
-    );
-
-    const actionButtonsDOM = newQuery && !isStreaming && (
-        <div className="right-align mt12">
-            <Button title="Accept" onClick={handleAccept} color="confirm" />
-            <Button title="Discard" onClick={handleDiscard} />
-        </div>
-    );
-
-    return (
-        <div>
-            {tablesDOM}
-            {explanation && <div className="mt12">{explanation}</div>}
-            {queryDiffDOM}
-            {actionButtonsDOM}
-        </div>
-    );
-};
-
 export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
     ({ query = '', queryEngine, onUpdateQuery, onFormatQuery }, ref) => {
         const defaultCommand = QUERY_CELL_COMMANDS.find(
@@ -199,15 +67,19 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
         const [showPopupView, setShowPopupView] = useState(false);
         const [command, setCommand] =
             useState<IQueryCellCommand>(defaultCommand);
-        const [commandInputValue, setCommandInputValue] = useState<string>();
+        const [commandInputValue, setCommandInputValue] = useState<string>('');
         const [commandRunner, setCommandRunner] = useState<CommandRunner>();
         const [commandKwargs, setCommandKwargs] = useState<Record<string, any>>(
             {}
         );
-        const [resetResult, setResetResult] = useState(false);
 
-        const { runCommand, isRunning, cancelCommand, commandResult } =
-            useCommand(command, commandRunner, resetResult);
+        const {
+            runCommand,
+            isRunning,
+            cancelCommand,
+            commandResult,
+            resetCommandResult,
+        } = useCommand(command, commandRunner);
 
         useEffect(() => {
             setTables((tables) => uniq([...tablesInQuery, ...tables]));
@@ -236,7 +108,7 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
 
         const handleCommand = useCallback(() => {
             runCommand(commandKwargs);
-            setResetResult(false);
+            resetCommandResult();
             setShowPopupView(!command.inplace);
 
             if (command.name === 'generate' || command.name === 'edit') {
@@ -262,7 +134,7 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
         );
         useEvent('keydown', onEscapeKeyDown);
 
-        const getCommandResultView = useCallback(() => {
+        const getCommandResultView = () => {
             if (!commandResult) {
                 return null;
             }
@@ -282,11 +154,11 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
                         onAccept={(query) => {
                             onUpdateQuery(query, false);
                             setShowPopupView(false);
-                            setResetResult(true);
+                            resetCommandResult();
                         }}
                         onDiscard={() => {
                             setShowPopupView(false);
-                            setResetResult(true);
+                            resetCommandResult();
                         }}
                     />
                 );
@@ -294,14 +166,7 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
 
             // TODO: Handle other command types, or have it covered by AICommandResultView
             return null;
-        }, [
-            commandResult,
-            queryEngine,
-            tables,
-            query,
-            isRunning,
-            handleCommand,
-        ]);
+        };
 
         return (
             <div
@@ -316,18 +181,20 @@ export const AICommandBar: React.FC<IQueryCellCommandBarProps> = forwardRef(
                 <div
                     className={showPopupView ? 'command-popup-view' : undefined}
                 >
-                    {showPopupView && (
-                        <Message
-                            message="Note: This AI-powered query generation may be inaccurate. Please use your own judgement and verify the result."
-                            type="warning"
-                            className="warning-message"
-                        />
-                    )}
+                    {showPopupView &&
+                        command.type === QueryCellCommandType.AI && (
+                            <Message
+                                message="Note: This AI-powered query generation may be inaccurate. Please use your own judgement and verify the result."
+                                type="warning"
+                                className="warning-message"
+                            />
+                        )}
                     <AICommandInput
                         commands={QUERY_CELL_COMMANDS}
-                        placeholder={command.hint}
                         onCommandChange={(command, inputValue) => {
-                            setCommand(command || defaultCommand);
+                            setCommand(
+                                (oldCommand) => command ?? defaultCommand
+                            );
                             setCommandInputValue(inputValue);
                         }}
                         onSubmit={handleCommand}
